services/image_analysis/requirements.txt
fastapi
uvicorn
pydantic
python-multipart==0.0.6
 Pillow
numpy